Allow reconstruction labels to disappear
This commit removes the constraint that each reconstruction label has to
be used at least once in the corrected reconstruction. This caused
unintuitive results, where small fragments that could not be explained
away due to this constraint caused a chain reaction resulting in
hard-to-interpret error report. Furthermore, following the analogy of
elastically deforming the reconstruction within tolerance bounds, the
disappearance of small fragments is acceptable, as long as they are
smaller than the threshold.